* Not being able to lock the config file on linux is not a fatal error anymore   (close mantis: #676) * UnitDef::flankingBonusMode was read as float instead of int * Refactored CGameSetup, should be somewhat better readable now. (mostly split   the huge Init function in a lot of smaller ones, and increased locality) * Made LuaGaia and LuaRules tags work in startscript, key & default value   arguments where swapped.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@4819 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Units/UnitDefHandler.cpp
@@ -306,7 +306,7 @@ void CUnitDefHandler::ParseTAUnit(const LuaTable& udTable, const string& unitNam
 	ud.captureSpeed   = udTable.GetFloat("captureSpeed",   ud.buildSpeed);
 	ud.terraformSpeed = udTable.GetFloat("terraformSpeed", ud.buildSpeed);
 
-	ud.flankingBonusMode = udTable.GetFloat("flankingBonusMode", modInfo.flankingBonusModeDefault);
+	ud.flankingBonusMode = udTable.GetInt("flankingBonusMode", modInfo.flankingBonusModeDefault);
 	ud.flankingBonusMax  = udTable.GetFloat("flankingBonusMax", 1.9f);
 	ud.flankingBonusMin  = udTable.GetFloat("flankingBonusMin", 0.9);
 	ud.flankingBonusDir  = udTable.GetFloat3("flankingBonusDir", float3(0.0f, 0.0f, 1.0f));
